add comment and filter directory

+++ src/th/co/main/Application.java
@@ -16,31 +16,39 @@ public class Application {
 		String targetDir = "D:/test/";
 		//set destination path
 		String outputPath = "D:/output/";
+		//set dictionary path
+		String dictionaryPath = "D:/test/dictionary.txt";
 		File dir = new File(targetDir);
 		File[] allFileInDir = dir.listFiles();
 		//read dictionary
-		HashMap<String, String> replaceWord  = readDict("D:/test/dictionary.txt");
+		HashMap<String, String> replaceWord  = readDict(dictionaryPath);
 		//start process each file
 		for(File file : allFileInDir){
-			FileReader reader = new FileReader(file);
-			BufferedReader br = new BufferedReader(reader);
-			FileWriter writer = new FileWriter(new File(outputPath + file.getName()));
-			BufferedWriter bw = new BufferedWriter(writer);
-			String line;
-			//read each file
-			while((line = br.readLine()) != null){
-				System.out.println(line);
-				//each key word in dictionary
-				for(String key : replaceWord.keySet()){
-					//replace text
-					line = line.replace(key, replaceWord.get(key));
+			if(file.isFile()){
+				System.out.println(file.getName() + " is File start process");
+				FileReader reader = new FileReader(file);
+				BufferedReader br = new BufferedReader(reader);
+				FileWriter writer = new FileWriter(new File(outputPath + file.getName()));
+				BufferedWriter bw = new BufferedWriter(writer);
+				String line;
+				//read each file
+				while((line = br.readLine()) != null){
+					System.out.println("old line : " + line);
+					//each key word in dictionary
+					for(String key : replaceWord.keySet()){
+						//replace text
+						line = line.replace(key, replaceWord.get(key));
+					}
+					//write processed line to file
+					bw.write(line);
+					bw.newLine();
+					bw.flush();
+					System.out.println("new line : " + line);
 				}
-				bw.write(line);
-				bw.newLine();
-				bw.flush();
-				System.out.println(line);
+				bw.close();
+			}else{
+				System.out.println(file.getName() + " is directory not process");
 			}
-			bw.close();
 		}
 
 	}
@@ -54,6 +62,7 @@ public class Application {
 		while((line = br.readLine()) != null){
 			String[] s;
 			s = line.split("=");
+			//expect format oldValue=newValue
 			dictionary.put(s[0], s[1]);
 		}
 		return dictionary;
